+++ src/main/java/net/creeperhost/chickens/blockentities/BlockEntityRoost.java
@@ -44,7 +44,7 @@ import java.util.Random;
 
 public class BlockEntityRoost extends BlockEntity implements MenuProvider
 {
-    public SmartInventory inventory = new SmartInventory(6)
+    public SmartInventory inventory = new SmartInventory(5)
     {
         @Override
         protected void onContentsChanged(int slot)
@@ -137,7 +137,7 @@ public class BlockEntityRoost extends BlockEntity implements MenuProvider
                 }
                 else
                 {
-                    ItemStack inserted = moveOutput(getOutputStack(), false);//InventoryHelper.insertItemStacked(inventory, getOutputStack(), false);
+                    ItemStack inserted = InventoryHelper.insertItemStacked(inventory, getOutputStack(), false);
                     if(inserted.isEmpty())
                     {
                         level.playSound(null, getBlockPos(), SoundEvents.CHICKEN_EGG, SoundSource.NEUTRAL, 0.5F, 0.8F);
@@ -162,57 +162,11 @@ public class BlockEntityRoost extends BlockEntity implements MenuProvider
         return Math.min(value, 50);
     }
 
-    //TODO I hate all of this but its a quick fix
-    public ItemStack moveOutput(ItemStack stack, boolean simulate)
-    {
-        for (int i = 1; i <= 4; i++)
-        {
-            if(inventory.getStackInSlot(i).isEmpty())
-            {
-                if(!simulate) inventory.setStackInSlot(i, stack);
-                return ItemStack.EMPTY;
-            }
-            else
-            {
-                if(ItemStack.isSameItemSameTags(stack, inventory.getStackInSlot(i)))
-                {
-                    int startStackSize = stack.getCount();
-                    int count = inventory.getStackInSlot(i).getCount();
-                    int max = stack.getMaxStackSize();
-                    if(count <= max)
-                    {
-                        if(startStackSize + count > max)
-                        {
-                            int fitSize = max - count;
-                            if(fitSize > startStackSize)
-                            {
-                                int returnSize = startStackSize - fitSize;
-                                stack.setCount(fitSize);
-                                if(!simulate) inventory.setStackInSlot(i, stack);
-                                ItemStack returnStack = stack.copy();
-                                returnStack.setCount(returnSize);
-                                return returnStack;
-                            }
-                        }
-                        else
-                        {
-                            int newCount = count + stack.getCount();
-                            stack.setCount(newCount);
-                            if (!simulate) inventory.setStackInSlot(i, stack);
-                            return ItemStack.EMPTY;
-                        }
-                    }
-                }
-            }
-        }
-        return stack;
-    }
-
     public boolean canRun()
     {
         if(inventory.getStackInSlot(0).isEmpty()) return false;
         if(!(inventory.getStackInSlot(0).getItem() instanceof ItemChicken)) return false;
-        ItemStack insert = moveOutput(getOutputStack(), true);//InventoryHelper.insertItemStacked(inventory, getOutputStack(), true);
+        ItemStack insert = InventoryHelper.insertItemStacked(inventory, getOutputStack(), true);
         return insert.isEmpty();
     }